Update 2.6.2
Changed RankManager.reloadAllPlayerRanks() function to get the
playerlist from the server, and not the config file.
Changed getOfflinePlayer methods to use UUIDs.
Moved "rm.reloadRanks()" in Main.java from Line 30 to Line 34, because
the debug_prefix has to be loaded before this function is called.
Updated version number in plugin.yml